Reorganize files into reproject.mosaicking sub-package
